flash/nor/stm32f1x: fix format specifiers warning
Caught by clang on OS X. Reported by Steffanx on IRC.

Reviewed-on: http://openocd.zylin.com/1879
Tested-by: jenkins

+++ src/flash/nor/stm32f1x.c
@@ -1347,9 +1347,9 @@ COMMAND_HANDLER(stm32x_handle_options_read_command)
 	}
 
 	command_print(CMD_CTX, "User Option0: 0x%02" PRIx8,
-			(user_data >> stm32x_info->user_data_offset) & 0xff);
+			(uint8_t)((user_data >> stm32x_info->user_data_offset) & 0xff));
 	command_print(CMD_CTX, "User Option1: 0x%02" PRIx8,
-			(user_data >> (stm32x_info->user_data_offset + 8)) & 0xff);
+			(uint8_t)((user_data >> (stm32x_info->user_data_offset + 8)) & 0xff));
 
 	return ERROR_OK;
 }
